+++ data/meta.sql
@@ -1,0 +1,6 @@
+-- 5 rows, ordered by key
+INSERT INTO meta(key, value) VALUES('dolibarr_root', '/home/nzaou/dolibarr');
+INSERT INTO meta(key, value) VALUES('extract_date', '2026-04-18T14:53:51.544953+00:00');
+INSERT INTO meta(key, value) VALUES('hash_version', 'v2');
+INSERT INTO meta(key, value) VALUES('prompt_version', 'v2');
+INSERT INTO meta(key, value) VALUES('seed_rows', '41');